backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.32
alembic==1.13.2
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
psycopg2-binary
redis==5.0.7
python-multipart==0.0.9
